fabric: Convert fi_gettime_ms to uint64_t
Convert fi_gettime_ms from a double to uint64_t to avoid
floating point operations.

+++ include/fi.h
@@ -207,7 +207,7 @@ int fi_recv_allowed(uint64_t caps);
 int fi_rma_initiate_allowed(uint64_t caps);
 int fi_rma_target_allowed(uint64_t caps);
 
-double fi_gettime_ms();
+uint64_t fi_gettime_ms();
 
 #define RDMA_CONF_DIR  SYSCONFDIR "/" RDMADIR
 #define FI_CONF_DIR RDMA_CONF_DIR "/fabric"
